requirement.txt
mcstatus
request
flask